BackEnd/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
transformers==4.35.0
torch==2.9.0
python-dotenv==1.0.0
